+++ MauiClient/Resources/Constants/Constants.xaml.cs
@@ -1,0 +1,9 @@
+namespace MauiClient.Resources.Constants;
+
+public partial class Constants : ResourceDictionary
+{
+	public Constants()
+	{
+		InitializeComponent();
+	}
+}